requirements.txt
faiss-cpu
sentence_transformers==2.2.2
langchain
langchain_community
zhipuai
gradio
jieba
pymilvus
pypinyin

pandas